New class and function moved to a separate module, pc_fallback.py
