+++ .bash_history
@@ -56,3 +56,7 @@ nano laalaa/apps/advisers/geocoder.py
 ./runserver.sh 
 exit
 ./runonce.sh 
+./runserver.sh 
+export POSTCODEINFO_AUTH_TOKEN=a4a3c5ed091552daf8d4a6b87a43bf100e359e89
+./runserver.sh 
+exit
