+++ laalaa/apps/advisers/geocoder.py
@@ -1,38 +1,16 @@
 import re
 import postcodeinfo
-from postcodes import PostCoder
+import pc_fallback
 
 
 class GeocoderError(Exception):
     pass
 
 
-class PostcodePlaceholder:
-
-    def __init__(self):
-        self.longitude = None
-        self.latitude = None
-
-
 class PostcodeNotFound(GeocoderError):
     def __init__(self, postcode, *args, **kwargs):
         super(PostcodeNotFound, self).__init__(*args, **kwargs)
         self.postcode = postcode
-
-
-def geocode_fallback(postcode):
-
-    result = PostcodePlaceholder()
-
-    print '******CHECKING ' + postcode + 'AT POSTCODES.IO********'
-    postcode_data = PostCoder().get(postcode)
-    print '******RESPONSE:'
-    print result
-
-    result.latitude = postcode_data['geo']['lat']
-    result.longitude = postcode_data['geo']['lng']
-
-    return result
 
 
 def geocode(postcode):
@@ -44,7 +22,7 @@ def geocode(postcode):
 
         if not result.valid:
             print '******Result not valid********'
-            result = geocode_fallback(postcode)
+            result = pc_fallback.geocode(postcode)
 
             if not result.latitude:
                 raise PostcodeNotFound(postcode)
